Split the heavy LAS bridge so the eager shell stays light
The bridge statically imported the whole out-of-core cluster into
openLocalHeavyLas, and openScan imports that eagerly, so storagePreflight,
tileStoreBuilder, opfsSpillStore, OlvTileSource, the tile decoder and the worker
glue all landed in the index chunk and pushed it 24 KiB over budget.

The decision half now keeps only the capability probe, the header peek and
planLoad through the buildThenStream gate, all light imports. Everything from
the preflight onward moves to heavyLasExecutor, reached through a new
lazyChunks.loadHeavyLasExecutor dynamic import that fires only after the plan
routes a file out of core. Shared shapes move to heavyLasTypes. A small LAS, any
LAZ, or a non-capable environment never loads the cluster. Index is back under
780 KiB and the fail-safe behaviour and every test are unchanged.
